update test api_signin
