+++ message_app/model.py
@@ -1,0 +1,17 @@
+import sqlite3
+from message_app import create_app
+from flask import g
+
+DATABASE = 'VL_MESSAGES.db'
+
+def get_db():
+    db = getattr(g, '_database', None)
+    if db is None:
+        db = g._database = sqlite3.connect(DATABASE)
+    return db
+
+@app.teardown_appcontext
+def close_connection(exception):
+    db = getattr(g, '_database', None)
+    if db is not None:
+        db.close()
