--- message_app/tests/conftest.py
@@ -1,5 +1,0 @@
-import pytest
-
-@pytest.fixture
-def supply_url():
-	return "http://localhost:5000/api"
